+++ server/usecase/playerUsecase.ts
@@ -4,13 +4,13 @@ import {
   DISPLAY_COUNT,
   SCREEN_HEIGHT,
   SCREEN_WIDTH,
-  SPEED_BOOST_DURATION_MS,
 } from '../commonConstantsWithClient';
 import type { UserId } from '../commonTypesWithClient/branded';
 import type { PlayerModel } from '../commonTypesWithClient/models';
 import { playerRepository } from '../repository/playerRepository';
 import { computeAllowedMoveX } from '../service/computeAllowedMoveX';
 import { userIdParser } from '../service/idParsers';
+import { itemHandler } from '../service/itemhandler';
 import { minMax } from '../service/minMax';
 import { itemsData, type Item } from './../commonConstantsWithClient/item';
 
@@ -69,40 +69,6 @@ export const playerUseCase = {
     return await playerRepository.save(updatePlayerInfo);
   },
   useitem: async (userId: UserId, items: Item[]) => {
-    interface ItemHandlers {
-      [key: string]: (player: PlayerModel) => Promise<void>;
-    }
-    const itemHandler: ItemHandlers = {
-      speed: async (player: PlayerModel) => {
-        try {
-          const updatePlayerInfo: PlayerModel = {
-            ...player,
-            speed: DEFAULT_PLAYER_MOVE_SPEED * 2,
-            Items: (player.Items ?? []).slice(1),
-          };
-          console.log('speed up', updatePlayerInfo.speed);
-          await playerRepository.save(updatePlayerInfo);
-
-          setTimeout(async () => {
-            try {
-              const currentPlayer = await playerRepository.find(player.id);
-              if (currentPlayer === null) return null;
-              const updatePlayerInfo: PlayerModel = {
-                ...currentPlayer,
-                speed: DEFAULT_PLAYER_MOVE_SPEED,
-              };
-              console.log('speed down');
-              await playerRepository.save(updatePlayerInfo);
-            } catch (error) {
-              console.error('速度のリセットに失敗しました:', error);
-            }
-          }, SPEED_BOOST_DURATION_MS);
-        } catch (error) {
-          console.error('スピードアップに失敗しました:', error);
-        }
-      },
-    };
-
     const currentPlayer = await playerRepository.find(userId);
     if (currentPlayer === null) return null;
     const handler = itemHandler[items[0].name];
